Write function to display task card on saveTask()
Create task-card-display.js which stores all functions related to
manipulating the task card display in the DOM. Create a function that
takes each property of a task object as its parameter and generates
a task card using this information. At this point, the task card is
unstyled and lacks edit/delete buttons.