Move "save" and "cancel" to edit/cancel handler
Combine all possible buttons into the same event delegation handler.

Write save button functionality: run a submit button handler that
returns the input value, then use that input value as a parameter in a
function that updates the card content and in a function that updates
its corresponding project in local storage.

Delete button functionality: removes the edit form and generates the
card content with the preexisting title.